LPS-132020 Move util component to util package
