+++ modules/apps/data-cleanup/data-cleanup/src/main/java/com/liferay/data/cleanup/internal/upgrade/util/ConfigurationPersistenceManagerUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.data.cleanup.internal.upgrade;
+package com.liferay.data.cleanup.internal.upgrade.util;
 
 import aQute.bnd.annotation.metatype.Meta;
